Use jline / jansi to have the progress bars work on Windows too

+++ scripts/generate-graalvm-launcher.sh
@@ -3,28 +3,65 @@ set -eu
 
 BASE="$(dirname "${BASH_SOURCE[0]}")"
 
-ASSEMBLY="coursier-assembly"
+TRANSIENT_ASSEMBLY=false
+
+if [[ -z "${ASSEMBLY:-}" ]]; then
+  # TRANSIENT_ASSEMBLY=true
+  ASSEMBLY="coursier-assembly"
+
+  # Set that one too once it's published
+  # MODULE="coursier-cli-graalvm"
+  DOWNLOAD_LAUNCHER=true OUTPUT="$ASSEMBLY" "$BASE/generate-launcher.sh" --assembly "$@"
+fi
 
 cleanup() {
-  rm -f "$ASSEMBLY"
+  [[ "$TRANSIENT_ASSEMBLY" == false ]] || rm -f "$ASSEMBLY"
 }
 
 trap cleanup EXIT INT TERM
 
 
-OUTPUT="$ASSEMBLY" "$BASE/generate-launcher.sh" --assembly
+if [[ "${TEST_RUN:-false}" == true ]]; then
+  echo "Test run, not generating graalvm image" 1>&2
+  exit 0
+fi
 
-OUTPUT="coursier-graalvm"
+OUTPUT="${OUTPUT:-"coursier-graalvm"}"
 
-"$GRAALVM_HOME/bin/native-image" \
+NATIVE_IMAGE="${NATIVE_IMAGE:-"$GRAALVM_HOME/bin/native-image"}"
+
+JAVA_SECURITY_PROPS="$BASE/java.security.overrides"
+
+if [[ ! -f "$JAVA_SECURITY_PROPS" ]]; then
+  echo "$JAVA_SECURITY_PROPS not found." 1>&2
+  exit 1
+fi
+
+"$NATIVE_IMAGE" \
+  --no-server \
   --enable-http \
   --enable-https \
-  -H:ReflectionConfigurationFiles="$BASE/reflection.json" \
   --no-fallback \
-  --no-server \
+  -H:EnableURLProtocols=http,https \
+  --enable-all-security-services \
+  -H:ReflectionConfigurationFiles="$BASE/reflection.json" \
+  -H:+ReportExceptionStackTraces \
+  --initialize-at-build-time=scala.Symbol \
+  --initialize-at-build-time=scala.Function2 \
+  --initialize-at-build-time=scala.runtime.StructuralCallSite \
+  --initialize-at-build-time=scala.runtime.EmptyMethodCache \
+  --initialize-at-build-time=coursier.util.Properties\$ \
+  --initialize-at-build-time=org.jline.utils.InputStreamReader \
+  --initialize-at-build-time=org.jline.utils.InfoCmp \
+  --initialize-at-build-time=coursier.cli.CoursierGraalvm \
+  --allow-incomplete-classpath \
+  -J-D"java.security.properties=$JAVA_SECURITY_PROPS" \
+  -D"java.security.properties=$JAVA_SECURITY_PROPS" \
   -jar "$ASSEMBLY" \
   "$OUTPUT"
 
+if [ -e "$OUTPUT.exe" -a ! -e "$OUTPUT" ]; then
+  mv "$OUTPUT.exe" "$OUTPUT"
+fi
+
 echo "Generate graalvm native image $OUTPUT."
-echo "Run like"
-echo "  coursier-graalvm -Djava.library.path=\$GRAALVM_HOME/jre/lib …args…"
